# [WIP] Optimizations for actor creation