**WORK IN PROGRESS**

I'm working on some optimizations around creating a child actors. Those will include:

- Adding `sealed` modifier where possible (current .NET Core compiler can inline virtcall for sealed classes).
- Replacing interfaces with actual classes where possible (again, avoid virtcalls, and add possibility to inline).
- Replacing scala-like interface-sealed class with tagged structs where this has any sense. Reasons:
    - Maybe JVM is able to specialize case class switch to more optimal version, but I guess struct + int-based switches and comparison is faster in .NET.
    - Also structs &rarr; no heap allocations.

Additionally I'm thinking about replacing internal sets used by `IChildrenContainer` implementations from immutable to mutable versions. **Is there any reason to actually use immutable collections?**

Some benchmarks comparing immutable/mutable set operations:

```
BenchmarkDotNet=v0.10.11, OS=Windows 10 Redstone 3 [1709, Fall Creators Update] (10.0.16299.125)
Processor=Intel Core i5-6300HQ CPU 2.30GHz (Skylake), ProcessorCount=4
Frequency=2250000 Hz, Resolution=444.4444 ns, Timer=TSC
.NET Core SDK=2.1.2
  [Host]     : .NET Core 2.0.3 (Framework 4.6.25815.02), 64bit RyuJIT
  DefaultJob : .NET Core 2.0.3 (Framework 4.6.25815.02), 64bit RyuJIT


                       Method |         Mean |      Error |     StdDev |          Min |          Max |         Op/s |  Gen 0 | Allocated |
----------------------------- |-------------:|-----------:|-----------:|-------------:|-------------:|-------------:|-------:|----------:|
          ImmutableSetAddItem |  1,887.23 ns | 25.5842 ns | 23.9314 ns |  1,859.94 ns |  1,935.84 ns |    529,877.2 | 0.1545 |     488 B |
            MutableSetAddItem |     34.31 ns |  0.1323 ns |  0.1033 ns |     33.98 ns |     34.35 ns | 29,146,988.5 |      - |       0 B |
    SortedImmutableSetAddItem |    753.85 ns |  0.6408 ns |  0.4634 ns |    753.26 ns |    754.67 ns |  1,326,527.6 | 0.1469 |     464 B |
      SortedMutableSetAddItem |    113.08 ns |  0.5415 ns |  0.4800 ns |    112.26 ns |    113.73 ns |  8,843,004.9 |      - |       0 B |
       ImmutableSetRemoveItem |  1,179.49 ns | 22.7134 ns | 25.2459 ns |  1,149.36 ns |  1,234.43 ns |    847,821.2 | 0.1183 |     376 B |
         MutableSetRemoveItem |     26.66 ns |  0.3931 ns |  0.3677 ns |     26.30 ns |     27.19 ns | 37,503,546.6 |      - |       0 B |
 SortedImmutableSetRemoveItem |    643.20 ns | 12.5951 ns | 11.7815 ns |    627.83 ns |    672.95 ns |  1,554,729.5 | 0.1163 |     368 B |
   SortedMutableSetRemoveItem |    198.63 ns |  0.3239 ns |  0.2529 ns |    198.17 ns |    199.02 ns |  5,034,478.1 |      - |       0 B |
        ImmutableSetEnumerate | 76,902.77 ns | 64.8008 ns | 57.4442 ns | 76,796.23 ns | 76,986.83 ns |     13,003.4 |      - |       0 B |
          MutableSetEnumerate |  1,728.08 ns |  1.1556 ns |  1.0809 ns |  1,726.34 ns |  1,729.49 ns |    578,677.4 |      - |       0 B |
  SortedImmutableSetEnumerate | 34,667.19 ns | 14.5656 ns | 13.6247 ns | 34,644.73 ns | 34,692.45 ns |     28,845.7 |      - |       0 B |
    SortedMutableSetEnumerate | 12,944.08 ns | 24.1978 ns | 22.6347 ns | 12,919.03 ns | 13,000.30 ns |     77,255.4 | 0.0458 |     192 B |

                              Method |          Mean |         Error |        StdDev |           Min |           Max |         Op/s |  Gen 0 | Allocated |
------------------------------------ |--------------:|--------------:|--------------:|--------------:|--------------:|-------------:|-------:|----------:|
          ImmutableDictionaryAddItem |   3,248.60 ns |    61.8991 ns |    57.9005 ns |   3,185.30 ns |   3,407.38 ns |    307,824.7 | 0.1945 |     616 B |
            MutableDictionaryAddItem |      48.03 ns |     0.2393 ns |     0.1998 ns |      47.74 ns |      48.50 ns | 20,821,790.1 |      - |       0 B |
    SortedImmutableDictionaryAddItem |   4,019.00 ns |    31.5921 ns |    28.0056 ns |   3,999.38 ns |   4,085.58 ns |    248,818.1 | 0.1602 |     528 B |
      SortedMutableDictionaryAddItem |   3,048.74 ns |    13.6629 ns |    12.7803 ns |   3,033.29 ns |   3,078.06 ns |    328,004.3 |      - |       0 B |
       ImmutableDictionaryRemoveItem |   2,711.59 ns |    21.5189 ns |    19.0759 ns |   2,682.92 ns |   2,751.26 ns |    368,787.8 | 0.1488 |     472 B |
         MutableDictionaryRemoveItem |     210.24 ns |     0.6072 ns |     0.5680 ns |     209.56 ns |     211.27 ns |  4,756,369.0 | 0.0150 |      48 B |
 SortedImmutableDictionaryRemoveItem |   4,395.44 ns |     7.3320 ns |     6.8583 ns |   4,385.01 ns |   4,409.59 ns |    227,508.3 | 0.1297 |     432 B |
   SortedMutableDictionaryRemoveItem |   3,518.90 ns |    10.2988 ns |     9.6335 ns |   3,504.25 ns |   3,530.31 ns |    284,180.1 | 0.0114 |      48 B |
        ImmutableDictionaryEnumerate | 120,396.39 ns | 1,384.9060 ns | 1,295.4419 ns | 119,174.50 ns | 123,571.48 ns |      8,305.9 |      - |       0 B |
          MutableDictionaryEnumerate |   2,437.69 ns |    11.2547 ns |     9.9770 ns |   2,419.62 ns |   2,453.23 ns |    410,224.6 |      - |       0 B |
  SortedImmutableDictionaryEnumerate |  64,721.63 ns |   577.1861 ns |   511.6604 ns |  64,041.74 ns |  65,494.60 ns |     15,450.8 |      - |       0 B |
    SortedMutableDictionaryEnumerate |  13,735.89 ns |    96.4481 ns |    80.5385 ns |  13,638.55 ns |  13,870.67 ns |     72,802.0 | 0.0458 |     192 B |
````
The operations are:

- Add 1 element to set filled with 201 elements.
- Remove 1 element from set filled with 201 elements.
- Enumerate (using foreach) over set of 201 elements.

I didn't checked dictionaries yet, but I guess the improvement is very similar. Therefore, if there's no actual reason to use immutables, we probably shouldn't use them.